Changed the way max_lag_ms in query is computed
It now remembers if the query is waiting for long time

+++ include/query_processor.h
@@ -72,7 +72,7 @@ class Query_Processor_Output {
 	int sticky_conn;
 	int multiplex;
 	int gtid_from_hostgroup;
-	int max_lag_ms;
+	long long max_lag_ms;
 	int log;
 	char *comment; // #643
 	std::string *new_query;
